Move mB TypedDashboard to dashboard app
